+++ custom_components/kumo_cloud/__init__.py
@@ -15,7 +15,13 @@ from .coordinator import KumoCloudConfigEntry, KumoCloudDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
-PLATFORMS: list[Platform] = [Platform.BINARY_SENSOR, Platform.CLIMATE, Platform.SENSOR]
+PLATFORMS: list[Platform] = [
+    Platform.BINARY_SENSOR,
+    Platform.BUTTON,
+    Platform.CLIMATE,
+    Platform.SENSOR,
+    Platform.SWITCH,
+]
 
 async def async_setup_entry(hass: HomeAssistant, entry: KumoCloudConfigEntry) -> bool:
     """Set up Kumo Cloud from a config entry."""